backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dataclasses-json==0.5.7
Django==4.2.2
django-slack==5.19.0
djangorestframework==3.14.0
frozenlist==1.3.3
heyoo==0.0.9
idna==3.4
kombu==5.3.0
langchain==0.0.196
langchainplus-sdk==0.0.8
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
slack==0.0.2
slackclient==2.9.4
SQLAlchemy==2.0.15
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
tqdm==4.65.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
